test(apiserver): move the gRPC API tests into the package they exercise
The API's 27 tests lived in the server package and reached the API only
through its constructor, so server/apiserver itself had none. They need no
Server, so move them across as-is; the cache test now injects the one
connector type it creates instead of dex's whole connector registry, which
is what the injected config map is for.
